requirements.txt
fastapi
uvicorn
pillow
jinja2
python-multipart